back/pyproject.toml
[tool.poetry]
name = "fastapi-nuxt3-docker-boilerplate"
version = "0.1.0"
description = ""
authors = ["nt-shuree <shuree@nashatech.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["uvicorn"], version = "^0.111.1"}
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
gunicorn = "^22.0.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.3.4"
pytz = "^2024.1"
requests = "^2.32.3"
python-jose = "^3.3.0"
bcrypt = "^4.2.0"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.1.0"
isort = "^5.13.2"
pytest = "^8.3.2"
requests = "^2.32.3"
pre-commit = "^3.7.1"
autoflake = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
